feat: remove visual aspects of widget
the new plan is that sky ux will go ahead and make a first-class menu in
the omnibar. thus no ui is needed. just open the topic url in a new tab.

the only thing missing is a way to communicate to the new first-class
menu that the spa has updated the help key, thus the relelvant href in
the menu needs to be updated.