revert to pre 3.0.1-rc.* versions
large chuncks of the 3.0.1-rc.* line are no longer needed/wanted.
undoing in favor of starting over an beginning a new 3.0.1-rc.* line where
instead of removing logic, the legacy logic is maintained and only used
based on certain config props. the new open-in-new-tab logic will still
be present, but only used based on config.

large chunks of the 3.0.1-rc.* line will likely be copied over, but
structured differently in order to achieve the above.

this commit resets the unreleased changes so we can do a proper rc
branch since it appears this effor will take much longer.

achieved via:
    git checkout dc6ec8f --no-overlay -- .a